+++ freud/order.pyx
@@ -75,6 +75,8 @@ cdef class CubaticOrderParameter:
             orientation.
     """  # noqa: E501
     cdef freud._order.CubaticOrderParameter * thisptr
+    cdef n_replicates
+    cdef seed
 
     def __cinit__(self, t_initial, t_final, scale, n_replicates=1, seed=None):
         # run checks
@@ -105,6 +107,8 @@ cdef class CubaticOrderParameter:
         self.thisptr = new freud._order.CubaticOrderParameter(
             t_initial, t_final, scale, <float*> &r4[0, 0, 0, 0], n_replicates,
             seed)
+        self.n_replicates = n_replicates
+        self.seed = seed
 
     def compute(self, orientations):
         R"""Calculates the per-particle and global order parameter.
@@ -185,6 +189,18 @@ cdef class CubaticOrderParameter:
             self.thisptr.getGenR4Tensor().get()
         return np.asarray(gen_r4_tensor)
 
+    def __repr__(self):
+        return ("freud.order.{cls}(t_initial={t_initial}, t_final={t_final}, "
+                "scale={scale}, n_replicates={n_replicates}, "
+                "seed={seed})").format(cls=type(self).__name__,
+                                       t_initial=self.t_initial,
+                                       t_final=self.t_final,
+                                       scale=self.scale,
+                                       n_replicates=self.n_replicates,
+                                       seed=self.seed)
+
+    def __str__(self):
+        return repr(self)
 
 cdef class NematicOrderParameter:
     R"""Compute the nematic order parameter for a system of particles.
@@ -210,6 +226,7 @@ cdef class NematicOrderParameter:
             3x3 matrix corresponding to the average particle orientation.
     """  # noqa: E501
     cdef freud._order.NematicOrderParameter *thisptr
+    cdef u
 
     def __cinit__(self, u):
         # run checks
@@ -218,6 +235,8 @@ cdef class NematicOrderParameter:
 
         cdef vec3[float] l_u = vec3[float](u[0], u[1], u[2])
         self.thisptr = new freud._order.NematicOrderParameter(l_u)
+        self.u = freud.common.convert_array(
+            u, 1, dtype=np.float32, contiguous=True, array_name="u")
 
     def compute(self, orientations):
         R"""Calculates the per-particle and global order parameter.
@@ -261,6 +280,13 @@ cdef class NematicOrderParameter:
         cdef const float[:, ::1] nematic_tensor = \
             <float[:3, :3]> self.thisptr.getNematicTensor().get()
         return np.asarray(nematic_tensor)
+
+    def __repr__(self):
+        return "freud.order.{cls}(u={u})".format(cls=type(self).__name__,
+                                                 u=self.u.tolist())
+
+    def __str__(self):
+        return repr(self)
 
 
 cdef class HexOrderParameter:
@@ -366,6 +392,14 @@ cdef class HexOrderParameter:
         cdef unsigned int k = self.thisptr.getK()
         return k
 
+    def __repr__(self):
+        return "freud.order.{cls}(rmax={rmax}, k={k}, n={n})".format(
+            cls=type(self).__name__, rmax=self.rmax, k=self.K,
+            n=self.num_neigh)
+
+    def __str__(self):
+        return repr(self)
+
 
 cdef class TransOrderParameter:
     R"""Compute the translational order parameter for each particle.
@@ -387,6 +421,8 @@ cdef class TransOrderParameter:
             Box used in the calculation.
         num_particles (unsigned int):
             Number of particles.
+        K (float):
+            Normalization value (d_r is divided by K).
     """
     cdef freud._order.TransOrderParameter * thisptr
     cdef num_neigh
@@ -444,6 +480,19 @@ cdef class TransOrderParameter:
     def num_particles(self):
         cdef unsigned int np = self.thisptr.getNP()
         return np
+
+    @property
+    def K(self):
+        cdef float k = self.thisptr.getK()
+        return k
+
+    def __repr__(self):
+        return "freud.order.{cls}(rmax={rmax}, k={k}, n={n})".format(
+            cls=type(self).__name__, rmax=self.rmax, k=self.K,
+            n=self.num_neigh)
+
+    def __str__(self):
+        return repr(self)
 
 
 cdef class LocalQl:
@@ -516,12 +565,16 @@ cdef class LocalQl:
     cdef freud._order.LocalQl * qlptr
     cdef freud.box.Box m_box
     cdef rmax
+    cdef sph_l
+    cdef rmin
 
     def __cinit__(self, box, rmax, l, rmin=0, *args, **kwargs):
         cdef freud.box.Box b = freud.common.convert_box(box)
         if type(self) is LocalQl:
             self.m_box = b
             self.rmax = rmax
+            self.sph_l = l
+            self.rmin = rmin
             self.qlptr = new freud._order.LocalQl(
                 dereference(b.thisptr), rmax, l, rmin)
 
@@ -689,6 +742,17 @@ cdef class LocalQl:
         self.qlptr.computeAveNorm(<vec3[float]*> &l_points[0, 0], nP)
         return self
 
+    def __repr__(self):
+        return ("freud.order.{cls}(box={box}, rmax={rmax}, l={sph_l}, "
+                "rmin={rmin})").format(cls=type(self).__name__,
+                                       box=self.m_box,
+                                       rmax=self.rmax,
+                                       sph_l=self.sph_l,
+                                       rmin=self.rmin)
+
+    def __str__(self):
+        return repr(self)
+
 
 cdef class LocalQlNear(LocalQl):
     R"""A variant of the :class:`~LocalQl` class that performs its average
@@ -745,6 +809,7 @@ cdef class LocalQlNear(LocalQl):
                 dereference(b.thisptr), rmax, l, 0)
             self.m_box = b
             self.rmax = rmax
+            self.sph_l = l
             self.num_neigh = kn
 
     def __dealloc__(self):
@@ -810,6 +875,17 @@ cdef class LocalQlNear(LocalQl):
             self.m_box, points, points, self.num_neigh, nlist, True, self.rmax)
         cdef freud.locality.NeighborList nlist_ = defaulted_nlist[0]
         return super(LocalQlNear, self).computeAveNorm(points, nlist_)
+
+    def __repr__(self):
+        return ("freud.order.{cls}(box={box}, rmax={rmax}, l={sph_l}, "
+                "kn={kn})").format(cls=type(self).__name__,
+                                   box=self.m_box,
+                                   rmax=self.rmax,
+                                   sph_l=self.sph_l,
+                                   kn=self.num_neigh)
+
+    def __str__(self):
+        return repr(self)
 
 
 cdef class LocalWl(LocalQl):
@@ -897,6 +973,8 @@ cdef class LocalWl(LocalQl):
                 dereference(b.thisptr), rmax, l, rmin)
             self.m_box = b
             self.rmax = rmax
+            self.sph_l = l
+            self.rmin = rmin
 
     def __dealloc__(self):
         if type(self) is LocalWl:
@@ -943,6 +1021,17 @@ cdef class LocalWl(LocalQl):
         cdef np.complex64_t[::1] ave_norm_Wl = \
             <np.complex64_t[:n_particles]> self.thisptr.getAveNormWl().get()
         return np.asarray(ave_norm_Wl, dtype=np.complex64)
+
+    def __repr__(self):
+        return ("freud.order.{cls}(box={box}, rmax={rmax}, l={sph_l}, "
+                "rmin={rmin})").format(cls=type(self).__name__,
+                                       box=self.m_box,
+                                       rmax=self.rmax,
+                                       sph_l=self.sph_l,
+                                       rmin=self.rmin)
+
+    def __str__(self):
+        return repr(self)
 
 
 cdef class LocalWlNear(LocalWl):
@@ -997,6 +1086,7 @@ cdef class LocalWlNear(LocalWl):
                 dereference(b.thisptr), rmax, l, 0)
             self.m_box = b
             self.rmax = rmax
+            self.sph_l = l
             self.num_neigh = kn
 
     def __dealloc__(self):
@@ -1062,6 +1152,17 @@ cdef class LocalWlNear(LocalWl):
         cdef freud.locality.NeighborList nlist_ = defaulted_nlist[0]
         return super(LocalWlNear, self).computeAveNorm(points, nlist_)
 
+    def __repr__(self):
+        return ("freud.order.{cls}(box={box}, rmax={rmax}, l={sph_l}, "
+                "kn={kn})").format(cls=type(self).__name__,
+                                   box=self.m_box,
+                                   rmax=self.rmax,
+                                   sph_l=self.sph_l,
+                                   kn=self.num_neigh)
+
+    def __str__(self):
+        return repr(self)
+
 
 cdef class SolLiq:
     R"""Uses dot products of :math:`Q_{lm}` between particles for clustering.
@@ -1112,6 +1213,9 @@ cdef class SolLiq:
     cdef freud._order.SolLiq * thisptr
     cdef freud.box.Box m_box
     cdef rmax
+    cdef Qthreshold
+    cdef Sthreshold
+    cdef sph_l
 
     def __cinit__(self, box, rmax, Qthreshold, Sthreshold, l, *args, **kwargs):
         cdef freud.box.Box b = freud.common.convert_box(box)
@@ -1120,6 +1224,9 @@ cdef class SolLiq:
                 dereference(b.thisptr), rmax, Qthreshold, Sthreshold, l)
             self.m_box = b
             self.rmax = rmax
+            self.Qthreshold = Qthreshold
+            self.Sthreshold = Sthreshold
+            self.sph_l = l
 
     def __dealloc__(self):
         del self.thisptr
@@ -1260,6 +1367,19 @@ cdef class SolLiq:
         cdef unsigned int np = self.thisptr.getNP()
         return np
 
+    def __repr__(self):
+        return ("freud.order.{cls}(box={box}, rmax={rmax}, "
+                "Qthreshold={Qthreshold}, Sthreshold={Sthreshold}, "
+                "l={sph_l})").format(cls=type(self).__name__,
+                                     box=self.m_box,
+                                     rmax=self.rmax,
+                                     Qthreshold=self.Qthreshold,
+                                     Sthreshold=self.Sthreshold,
+                                     sph_l=self.sph_l)
+
+    def __str__(self):
+        return repr(self)
+
 
 cdef class SolLiqNear(SolLiq):
     R"""A variant of the :class:`~SolLiq` class that performs its average over nearest neighbor particles as determined by an instance of :class:`freud.locality.NeighborList`. The number of included neighbors is determined by the kn parameter to the constructor.
@@ -1318,6 +1438,9 @@ cdef class SolLiqNear(SolLiq):
                 dereference(b.thisptr), rmax, Qthreshold, Sthreshold, l)
             self.m_box = b
             self.rmax = rmax
+            self.Qthreshold = Qthreshold
+            self.Sthreshold = Sthreshold
+            self.sph_l = l
             self.num_neigh = kn
 
     def __dealloc__(self):
@@ -1368,6 +1491,20 @@ cdef class SolLiqNear(SolLiq):
             self.m_box, points, points, self.num_neigh, nlist, True, self.rmax)
         cdef freud.locality.NeighborList nlist_ = defaulted_nlist[0]
         return SolLiq.computeSolLiqNoNorm(self, points, nlist_)
+
+    def __repr__(self):
+        return ("freud.order.{cls}(box={box}, rmax={rmax}, "
+                "Qthreshold={Qthreshold}, Sthreshold={Sthreshold}, "
+                "l={sph_l}, kn={kn})").format(cls=type(self).__name__,
+                                              box=self.m_box,
+                                              rmax=self.rmax,
+                                              Qthreshold=self.Qthreshold,
+                                              Sthreshold=self.Sthreshold,
+                                              sph_l=self.sph_l,
+                                              kn=self.num_neigh)
+
+    def __str__(self):
+        return repr(self)
 
 
 cdef class RotationalAutocorrelation:
@@ -1469,3 +1606,10 @@ cdef class RotationalAutocorrelation:
     def azimuthal(self):
         cdef unsigned int azimuthal = self.thisptr.getL()
         return azimuthal
+
+    def __repr__(self):
+        return "freud.order.{cls}(l={sph_l})".format(cls=type(self).__name__,
+                                                     sph_l=self.l)
+
+    def __str__(self):
+        return repr(self)