+++ tests/test_box_Box.py
@@ -1,6 +1,6 @@
 import numpy as np
 import numpy.testing as npt
-from freud import box as bx
+import freud
 from collections import namedtuple
 import unittest
 import warnings
@@ -15,24 +15,24 @@ class TestBox(unittest.TestCase):
     def test_construct(self):
         """Test correct behavior for various constructor signatures"""
         with self.assertRaises(ValueError):
-            bx.Box()
+            freud.box.Box()
 
         with self.assertRaises(ValueError):
-            bx.Box(0, 0)
+            freud.box.Box(0, 0)
 
         with self.assertRaises(ValueError):
-            bx.Box(1, 2, is2D=False)
+            freud.box.Box(1, 2, is2D=False)
 
         with warnings.catch_warnings(record=True) as w:
             warnings.simplefilter("always")
-            bx.Box(1, 2, 3, is2D=True)
+            freud.box.Box(1, 2, 3, is2D=True)
             self.assertTrue(len(w) == 1)
 
-        box = bx.Box(1, 2)
+        box = freud.box.Box(1, 2)
         self.assertTrue(box.dimensions == 2)
 
     def test_get_length(self):
-        box = bx.Box(2, 4, 5, 1, 0, 0)
+        box = freud.box.Box(2, 4, 5, 1, 0, 0)
 
         npt.assert_allclose(box.Lx, 2, rtol=1e-6, err_msg="LxFail")
         npt.assert_allclose(box.Ly, 4, rtol=1e-6, err_msg="LyFail")
@@ -43,7 +43,7 @@ class TestBox(unittest.TestCase):
 
     def test_set_length(self):
         # Make sure we can change the lengths of the box after its creation
-        box = bx.Box(1, 2, 3, 1, 0, 0)
+        box = freud.box.Box(1, 2, 3, 1, 0, 0)
 
         box.Lx = 4
         box.Ly = 5
@@ -64,21 +64,21 @@ class TestBox(unittest.TestCase):
             box.L = [1, 2]
 
     def test_get_tilt_factor(self):
-        box = bx.Box(2, 2, 2, 1, 2, 3)
+        box = freud.box.Box(2, 2, 2, 1, 2, 3)
 
         npt.assert_allclose(box.xy, 1, rtol=1e-6, err_msg="TiltXYFail")
         npt.assert_allclose(box.xz, 2, rtol=1e-6, err_msg="TiltXZFail")
         npt.assert_allclose(box.yz, 3, rtol=1e-6, err_msg="TiltYZFail")
 
     def test_box_volume(self):
-        box3d = bx.Box(2, 2, 2, 1, 0, 0)
-        box2d = bx.Box(2, 2, 0, 0, 0, 0, is2D=True)
+        box3d = freud.box.Box(2, 2, 2, 1, 0, 0)
+        box2d = freud.box.Box(2, 2, 0, 0, 0, 0, is2D=True)
 
         npt.assert_allclose(box3d.volume, 8, rtol=1e-6, err_msg="Volume3DFail")
         npt.assert_allclose(box2d.volume, 4, rtol=1e-6, err_msg="Volume2DFail")
 
     def test_wrap_single_particle(self):
-        box = bx.Box(2, 2, 2, 1, 0, 0)
+        box = freud.box.Box(2, 2, 2, 1, 0, 0)
 
         testpoints = [0, -1, -1]
         npt.assert_allclose(box.wrap(testpoints)[0], -2, rtol=1e-6,
@@ -92,7 +92,7 @@ class TestBox(unittest.TestCase):
             box.wrap([1, 2])
 
     def test_wrap_multiple_particles(self):
-        box = bx.Box(2, 2, 2, 1, 0, 0)
+        box = freud.box.Box(2, 2, 2, 1, 0, 0)
 
         testpoints = [[0, -1, -1], [0, 0.5, 0]]
         npt.assert_allclose(box.wrap(testpoints)[0, 0], -2, rtol=1e-6,
@@ -103,7 +103,7 @@ class TestBox(unittest.TestCase):
                             err_msg="WrapFail")
 
     def test_wrap_multiple_images(self):
-        box = bx.Box(2, 2, 2, 1, 0, 0)
+        box = freud.box.Box(2, 2, 2, 1, 0, 0)
 
         testpoints = [[10, -5, -5], [0, 0.5, 0]]
         npt.assert_allclose(box.wrap(testpoints)[0, 0], -2, rtol=1e-6,
@@ -114,7 +114,7 @@ class TestBox(unittest.TestCase):
                             err_msg="WrapFail")
 
     def test_unwrap(self):
-        box = bx.Box(2, 2, 2, 1, 0, 0)
+        box = freud.box.Box(2, 2, 2, 1, 0, 0)
 
         testpoints = [0, -1, -1]
         imgs = [1, 0, 0]
@@ -138,7 +138,7 @@ class TestBox(unittest.TestCase):
             box.unwrap(testpoints[:, :2], imgs)
 
         # Now test 2D
-        box = bx.Box.square(1)
+        box = freud.box.Box.square(1)
 
         testpoints = [10, 0, 0]
         imgs = [10, 1, 2]
@@ -146,7 +146,7 @@ class TestBox(unittest.TestCase):
                             rtol=1e-6, err_msg="WrapFail")
 
     def test_images(self):
-        box = bx.Box(2, 2, 2, 0, 0, 0)
+        box = freud.box.Box(2, 2, 2, 0, 0, 0)
         testpoints = np.array([[50, 40, 30],
                                [-10, 0, 0]])
         testimages = np.array([box.getImage(vec) for vec in testpoints])
@@ -161,7 +161,7 @@ class TestBox(unittest.TestCase):
                          err_msg="ImageFail")
 
     def test_coordinates(self):
-        box = bx.Box(2, 2, 2)
+        box = freud.box.Box(2, 2, 2)
         f_point = np.array([[0.5, 0.25, 0.75],
                             [0, 0, 0],
                             [0.5, 0.5, 0.5]])
@@ -175,7 +175,7 @@ class TestBox(unittest.TestCase):
         npt.assert_equal(testcoordinates, point, err_msg="CoordinatesFail")
 
     def test_fraction(self):
-        box = bx.Box(2, 2, 2)
+        box = freud.box.Box(2, 2, 2)
         f_point = np.array([[0.5, 0.25, 0.75],
                             [0, 0, 0],
                             [0.5, 0.5, 0.5]])
@@ -192,7 +192,7 @@ class TestBox(unittest.TestCase):
         """Test getting lattice vectors"""
         b_list = [1, 2, 3, 0.1, 0.2, 0.3]
         Lx, Ly, Lz, xy, xz, yz = b_list
-        box = bx.Box.from_box(b_list)
+        box = freud.box.Box.from_box(b_list)
         self.assertEqual(
             box.getLatticeVector(0),
             [Lx, 0, 0]
@@ -207,7 +207,7 @@ class TestBox(unittest.TestCase):
         )
 
     def test_periodic(self):
-        box = bx.Box(1, 2, 3, 0, 0, 0)
+        box = freud.box.Box(1, 2, 3, 0, 0, 0)
         false = [False, False, False]
         true = [True, True, True]
         self.assertEqual(box.periodic, true)
@@ -235,75 +235,79 @@ class TestBox(unittest.TestCase):
         self.assertEqual(box.periodic, true)
 
     def test_equal(self):
-        box = bx.Box(2, 2, 2, 1, 0.5, 0.1)
-        box2 = bx.Box(2, 2, 2, 1, 0, 0)
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
+        box2 = freud.box.Box(2, 2, 2, 1, 0, 0)
         self.assertEqual(box, box)
         self.assertNotEqual(box, box2)
 
+    def test_repr(self):
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
+        self.assertEqual(box, eval(repr(box)))
+
     def test_str(self):
-        box = bx.Box(2, 2, 2, 1, 0.5, 0.1)
-        box2 = bx.Box(2, 2, 2, 1, 0.5, 0.1)
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
+        box2 = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
         self.assertEqual(str(box), str(box2))
 
     def test_to_dict(self):
         """Test converting box to dict"""
-        box = bx.Box(2, 2, 2, 1, 0.5, 0.1)
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
         box2 = box.to_dict()
         box_dict = {'Lx': 2, 'Ly': 2, 'Lz': 2, 'xy': 1, 'xz': 0.5, 'yz': 0.1}
         for k in box_dict:
             npt.assert_allclose(box_dict[k], box2[k])
 
     def test_tuple(self):
-        box = bx.Box(2, 2, 2, 1, 0.5, 0.1)
-        box2 = bx.Box.from_box(box.to_tuple())
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
+        box2 = freud.box.Box.from_box(box.to_tuple())
         self.assertEqual(box, box2)
 
     def test_from_box(self):
         """Test various methods of initializing a box"""
-        box = bx.Box(2, 2, 2, 1, 0.5, 0.1)
-        box2 = bx.Box.from_box(box)
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
+        box2 = freud.box.Box.from_box(box)
         self.assertEqual(box, box2)
 
         box_dict = {'Lx': 2, 'Ly': 2, 'Lz': 2, 'xy': 1, 'xz': 0.5, 'yz': 0.1}
-        box3 = bx.Box.from_box(box_dict)
+        box3 = freud.box.Box.from_box(box_dict)
         self.assertEqual(box, box3)
 
         with self.assertRaises(ValueError):
             box_dict['dimensions'] = 3
-            bx.Box.from_box(box_dict, 2)
+            freud.box.Box.from_box(box_dict, 2)
 
         BoxTuple = namedtuple('BoxTuple',
                               ['Lx', 'Ly', 'Lz', 'xy', 'xz', 'yz',
                                'dimensions'])
-        box4 = bx.Box.from_box(BoxTuple(2, 2, 2, 1, 0.5, 0.1, 3))
+        box4 = freud.box.Box.from_box(BoxTuple(2, 2, 2, 1, 0.5, 0.1, 3))
         self.assertEqual(box, box4)
 
         with self.assertRaises(ValueError):
-            bx.Box.from_box(BoxTuple(2, 2, 2, 1, 0.5, 0.1, 2), 3)
+            freud.box.Box.from_box(BoxTuple(2, 2, 2, 1, 0.5, 0.1, 2), 3)
 
-        box5 = bx.Box.from_box([2, 2, 2, 1, 0.5, 0.1])
+        box5 = freud.box.Box.from_box([2, 2, 2, 1, 0.5, 0.1])
         self.assertEqual(box, box5)
 
-        box6 = bx.Box.from_box(np.array([2, 2, 2, 1, 0.5, 0.1]))
+        box6 = freud.box.Box.from_box(np.array([2, 2, 2, 1, 0.5, 0.1]))
         self.assertEqual(box, box6)
 
         with self.assertRaises(ValueError):
-            bx.Box.from_box([2, 2, 2, 1, 0.5])
+            freud.box.Box.from_box([2, 2, 2, 1, 0.5])
 
-        box7 = bx.Box.from_matrix(box.to_matrix())
+        box7 = freud.box.Box.from_matrix(box.to_matrix())
         self.assertTrue(np.isclose(box.to_matrix(), box7.to_matrix()).all())
 
     def test_matrix(self):
-        box = bx.Box(2, 2, 2, 1, 0.5, 0.1)
-        box2 = bx.Box.from_matrix(box.to_matrix())
+        box = freud.box.Box(2, 2, 2, 1, 0.5, 0.1)
+        box2 = freud.box.Box.from_matrix(box.to_matrix())
         self.assertTrue(np.isclose(box.to_matrix(), box2.to_matrix()).all())
 
-        box3 = bx.Box(2, 2, 0, 0.5, 0, 0)
-        box4 = bx.Box.from_matrix(box3.to_matrix())
+        box3 = freud.box.Box(2, 2, 0, 0.5, 0, 0)
+        box4 = freud.box.Box.from_matrix(box3.to_matrix())
         self.assertTrue(np.isclose(box3.to_matrix(), box4.to_matrix()).all())
 
     def test_2_dimensional(self):
-        box = bx.Box.square(L=1)
+        box = freud.box.Box.square(L=1)
         # Setting Lz for a 2D box throws a warning that we hide with setUp()
         box.Lz = 1.0
         self.assertEqual(box.Lz, 0.0)
@@ -316,7 +320,7 @@ class TestBox(unittest.TestCase):
 
     def test_cube(self):
         L = 10.0
-        cube = bx.Box.cube(L=L)
+        cube = freud.box.Box.cube(L=L)
         self.assertEqual(cube.Lx, L)
         self.assertEqual(cube.Ly, L)
         self.assertEqual(cube.Lz, L)
@@ -327,7 +331,7 @@ class TestBox(unittest.TestCase):
 
     def test_square(self):
         L = 10.0
-        square = bx.Box.square(L=L)
+        square = freud.box.Box.square(L=L)
         self.assertEqual(square.Lx, L)
         self.assertEqual(square.Ly, L)
         self.assertEqual(square.Lz, 0)
@@ -337,7 +341,7 @@ class TestBox(unittest.TestCase):
         self.assertEqual(square.dimensions, 2)
 
     def test_multiply(self):
-        box = bx.Box(2, 3, 4, 1, 0.5, 0.1)
+        box = freud.box.Box(2, 3, 4, 1, 0.5, 0.1)
         box2 = box * 2
         self.assertTrue(np.isclose(box2.Lx, 4))
         self.assertTrue(np.isclose(box2.Ly, 6))