+++ tests/test_order_LocalQl.py
@@ -46,6 +46,11 @@ class TestLocalQl(unittest.TestCase):
 
         self.assertEqual(len(positions), comp.num_particles)
 
+    def test_repr(self):
+        box = freud.box.Box.cube(10)
+        comp = freud.order.LocalQl(box, 1.5, 6)
+        self.assertEqual(str(comp), str(eval(repr(comp))))
+
 
 class TestLocalQlNear(unittest.TestCase):
     def test_init_kwargs(self):
@@ -90,6 +95,11 @@ class TestLocalQlNear(unittest.TestCase):
         self.assertEqual(box, comp.box)
 
         self.assertEqual(len(positions), comp.num_particles)
+
+    def test_repr(self):
+        box = freud.box.Box.cube(10)
+        comp = freud.order.LocalQlNear(box, 1.5, 6, 12)
+        self.assertEqual(str(comp), str(eval(repr(comp))))
 
 
 if __name__ == '__main__':