+++ tests/test_order_TransOrderParameter.py
@@ -22,8 +22,11 @@ class TestTransOrder(unittest.TestCase):
         npt.assert_allclose(trans.d_r, 0, atol=1e-6)
 
         self.assertEqual(box, trans.box)
-
         self.assertEqual(len(positions), trans.num_particles)
+
+    def test_repr(self):
+        trans = freud.order.TransOrderParameter(1.1, 4, 4)
+        self.assertEqual(str(trans), str(eval(repr(trans))))
 
 
 if __name__ == '__main__':