+++ ChangeLog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 * New neighbor querying API to enable reuse of query data structures.
 * Exposed AABB tree to user.
+* All compute objects have `__repr__` and `__str__` methods defined.
 
 ### Changes
 * Use constant memoryviews to prevent errors with read-only inputs.